Move ignore_warnings decorator/context manager to astropy.tests.helper so that it can be used by more than just the FITS tests.
