+++ astropy/io/fits/tests/test_image.py
@@ -12,12 +12,11 @@ import numpy as np
 from ....io import fits
 from ....utils.exceptions import (AstropyDeprecationWarning,
                                   AstropyPendingDeprecationWarning)
-from ....tests.helper import pytest, raises, catch_warnings
+from ....tests.helper import pytest, raises, catch_warnings, ignore_warnings
 from ..hdu.compressed import SUBTRACTIVE_DITHER_1, DITHER_SEED_CHECKSUM
 from .test_table import comparerecords
 
 from . import FitsTestCase
-from .util import ignore_warnings
 
 
 class TestImageFunctions(FitsTestCase):
