--- astropy/io/fits/tests/util.py
@@ -1,45 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Test utility functions."""
-
-import functools
-import warnings
-
-from ....tests.helper import catch_warnings
-
-
-class ignore_warnings(catch_warnings):
-    """
-    This can be used either as a context manager or function decorator to
-    ignore all warnings that occur within a function or block of code.
-
-    An optional category option can be supplied to only ignore warnings of a
-    certain category or categories (if a list is provided).
-    """
-
-    def __init__(self, category=None):
-        super(ignore_warnings, self).__init__()
-
-        if isinstance(category, type) and issubclass(category, Warning):
-            self.category = [category]
-        else:
-            self.category = category
-
-    def __call__(self, func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            # Originally this just reused self, but that doesn't work if the
-            # function is called more than once so we need to make a new
-            # context manager instance for each call
-            with self.__class__(category=self.category):
-                return func(*args, **kwargs)
-
-        return wrapper
-
-    def __enter__(self):
-        retval = super(ignore_warnings, self).__enter__()
-        if self.category is not None:
-            for category in self.category:
-                warnings.simplefilter('ignore', category)
-        else:
-            warnings.simplefilter('ignore')
-        return retval
